Split out getPostPreviewUrl
